transfer rate is calculated with BigDecimal instead of float

+++ examples/src/main/java/fr/bmartel/speedtest/examples/DownloadFileExample.java
@@ -33,6 +33,8 @@ import fr.bmartel.speedtest.SpeedTestSocket;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.math.BigDecimal;
+
 /**
  * Download file Speed Test example.
  *
@@ -82,8 +84,8 @@ public class DownloadFileExample {
         speedTestSocket.addSpeedTestListener(new ISpeedTestListener() {
 
             @Override
-            public void onDownloadPacketsReceived(final long packetSize, final float transferRateBitPerSeconds, final
-            float transferRateOctetPerSeconds) {
+            public void onDownloadPacketsReceived(final long packetSize, final BigDecimal transferRateBitPerSeconds,
+                                                  final BigDecimal transferRateOctetPerSeconds) {
 
                 LogUtils.logFinishedTask(SpeedTestMode.DOWNLOAD, packetSize, transferRateBitPerSeconds,
                         transferRateOctetPerSeconds, LOGGER);
@@ -98,8 +100,8 @@ public class DownloadFileExample {
             }
 
             @Override
-            public void onUploadPacketsReceived(final long packetSize, final float transferRateBitPerSeconds, final
-            float transferRateOctetPerSeconds) {
+            public void onUploadPacketsReceived(final long packetSize, final BigDecimal transferRateBitPerSeconds, final
+            BigDecimal transferRateOctetPerSeconds) {
 
                 LogUtils.logFinishedTask(SpeedTestMode.UPLOAD, packetSize, transferRateBitPerSeconds,
                         transferRateOctetPerSeconds, LOGGER);
